rename lib to crypto
