+++ crypto/enclave_encrypt.go
@@ -1,4 +1,4 @@
-package lib
+package crypto
 
 import (
 	"bytes"
